Add 'payment' letter type and Mydata dir
Mydata directory is not part of repo (confidential info.)
I'm using it to contain addresses of entities to whom we pay for
services.  The following command:
'.utils.py prepare_mailing --which payment -i Mydata/vendors.csv'
is used to prepare a letter to accompany payments.

Still to do is set a lambda that can select the vendor we want.